Fix tests and flake8
