--- .github/workflows/releases.yml
@@ -1,46 +1,0 @@
-name: Releases
-
-on: [push]
-
-jobs:
-  release:
-    strategy:
-      matrix:
-        include:
-          - os: ubuntu-20.04
-            pyinstall_args: ""
-            wxpython: https://extras.wxpython.org/wxPython4/extras/linux/gtk2/ubuntu-18.04/wxPython-4.1.0-cp37-cp37m-linux_x86_64.whl
-          - os: ubuntu-18.04
-            pyinstall_args: ""
-            wxpython: https://extras.wxpython.org/wxPython4/extras/linux/gtk2/ubuntu-18.04/wxPython-4.1.0-cp37-cp37m-linux_x86_64.whl
-          - os: macos-10.15
-            pyinstall_args: --onefile --windowed
-            wxpython: wxpython
-          - os: windows-2019
-            pyinstall_args: --windowed
-            wxpython: wxpython
-
-    runs-on: ${{ matrix.os }}
-
-    steps:
-    - uses: actions/checkout@v2
-
-    - name: Set up Python 3.7
-      uses: actions/setup-python@v2
-      with:
-        python-version: 3.7
-
-    - name: Install Pip Dependencies
-      shell: bash
-      run: pip install pyinstaller pypubsub ${{ matrix.wxpython }}
-
-    - name: Build
-      shell: bash
-      run: pyinstaller ${{ matrix.pyinstall_args }} eplaunch.spec
-        # pyinstaller wf_tester.spec
-
-    - name: Upload
-      uses: actions/upload-artifact@v2
-      with:
-        name: EPLaunch-${{ matrix.os }}
-        path: dist/EPLaunch
