--- eplaunch/tests/interface/test_application.py
@@ -1,12 +1,0 @@
-import unittest
-
-# from eplaunch.interface.application import EpLaunchApplication
-
-
-class TestApplication(unittest.TestCase):
-
-    def test_initialization(self):
-        # ep_launch = EpLaunchApplication(None)
-        # init_response = ep_launch.OnInit()
-        # self.assertTrue(init_response)
-        pass
