--- eplaunch/tests/interface/test_frame.py
@@ -1,28 +1,0 @@
-import unittest
-
-# from eplaunch.interface.application import EpLaunchApplication
-
-
-class TestFrame(unittest.TestCase):
-
-    # def setUp(self):
-    #     app = EpLaunchApplication(0)
-    #     app.OnInit()
-    #     self.ep_launch_frame = app.frame_ep_launch
-
-    def test_any_threads_running(self):
-        """
-        Test that the frame threads running function responds to updates in the workflow_threads list
-
-        :return:
-        """
-        # self.ep_launch_frame.workflow_threads = []
-        # threads_running = self.ep_launch_frame.any_threads_running()
-        # self.assertFalse(threads_running)
-        # self.ep_launch_frame.workflow_threads = ["foo", "bar"]
-        # threads_running = self.ep_launch_frame.any_threads_running()
-        # self.assertTrue(threads_running)
-        # self.ep_launch_frame.workflow_threads = 2.0
-        # with self.assertRaises(TypeError):
-        #     self.ep_launch_frame.any_threads_running()
-        pass
